test(nextjs): Add MongoDB and modern SQL driver orchestrion instrumentations
Add e2e coverage for mongodb, mongoose, mysql2 and postgres.js orchestrion instrumentations in
the nextjs-16-orchestrion app. mongodb/mongoose use a new mongo container; mysql2 and postgres.js
reuse the existing MySQL and Postgres containers. `mysql2` is pinned below 3.20.0 (like `ioredis`)
so the orchestrion path is exercised rather than the driver's native diagnostics channels.

Also split the app's e2e tests into one file per instrumented library for readability.

Fixes #22505
