+++ dev-packages/e2e-tests/test-applications/nextjs-16-orchestrion/app/api/db-postgresjs/route.ts
@@ -1,0 +1,23 @@
+import { NextResponse } from 'next/server';
+import postgres from 'postgres';
+
+export const dynamic = 'force-dynamic';
+
+// postgres.js reuses the same Postgres container as the `pg` driver.
+export async function GET() {
+  const sql = postgres({
+    host: 'localhost',
+    port: 5432,
+    user: 'postgres',
+    password: 'docker',
+    database: 'postgres',
+  });
+
+  try {
+    await sql`SELECT 1 + 1 AS solution`;
+    await sql`SELECT * from generate_series(1, 3) as x`;
+    return NextResponse.json({ status: 'ok' });
+  } finally {
+    await sql.end();
+  }
+}
